# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "G4H-RMA Quant Engine",
    "description": "Enterprise Kalman + EGARCH + MCTS Pairs Trading with Multi-Agent System",
    "version": "5.0.0"
  },
  "paths": {
...
    "/api/v1/scan": {
      "post": {
        "summary": "Scan Pairs",
        "operationId": "scan_pairs_api_v1_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scan/{pair}": {
      "get": {
        "summary": "Scan Single Pair",
        "operationId": "scan_single_api_v1_scan__pair__get",
        "parameters": [
          {
            "name": "pair",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pair"
            }
          },
          {
            "name": "iterations",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 100,
              "default": 800,
              "title": "Iterations"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeSignal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/execute": {
      "post": {
        "summary": "Execute Signal",
        "operationId": "execute_signal_api_v1_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionType": {
        "type": "string",
        "enum": [
          "LONG_SPREAD",
          "SHORT_SPREAD",
          "HOLD"
        ],
        "title": "ActionType"
      },
      "AssetPair": {
        "properties": {
          "base": {
            "type": "string",
            "minLength": 1,
            "title": "Base"
          },
          "quote": {
            "type": "string",
            "minLength": 1,
            "title": "Quote"
          }
        },
        "type": "object",
        "required": [
          "base",
          "quote"
        ],
        "title": "AssetPair"
      },
      "EGARCHResult": {
        "properties": {
          "annualized_vol": {
            "type": "number",
            "minimum": 0.0,
            "title": "Annualized Vol"
          },
          "forecast_vol": {
            "type": "number",
            "minimum": 0.0,
            "title": "Forecast Vol"
          },
          "leverage_gamma": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leverage Gamma"
          },
          "regime": {
            "$ref": "#/components/schemas/VolatilityRegime"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "annualized_vol",
          "forecast_vol",
          "regime"
        ],
        "title": "EGARCHResult"
      },
      "ExecuteRequest": {
        "properties": {
          "base": {
            "type": "string",
            "title": "Base"
          },
          "quote": {
            "type": "string",
            "title": "Quote"
          },
          "action": {
            "$ref": "#/components/schemas/ActionType"
          },
          "qty": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qty"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "base",
          "quote",
          "action"
        ],
        "title": "ExecuteRequest"
      },
      "ExecutionResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "action": {
            "$ref": "#/components/schemas/ActionType"
          },
          "pair": {
            "type": "string",
            "title": "Pair"
          },
          "details": {
            "type": "string",
            "title": "Details",
            "default": ""
          },
          "order_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Order Ids"
          }
        },
        "type": "object",
        "required": [
          "status",
          "action",
          "pair"
        ],
        "title": "ExecutionResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KalmanState": {
        "properties": {
          "beta": {
            "type": "number",
            "title": "Beta"
          },
          "alpha": {
            "type": "number",
            "title": "Alpha"
          },
          "spread": {
            "type": "number",
            "title": "Spread"
          },
          "innovation_variance": {
            "type": "number",
            "title": "Innovation Variance"
          },
          "pure_z_score": {
            "type": "number",
            "title": "Pure Z Score"
          }
        },
        "type": "object",
        "required": [
          "beta",
          "alpha",
          "spread",
          "innovation_variance",
          "pure_z_score"
        ],
        "title": "KalmanState"
      },
      "MCTSResult": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionType"
          },
          "expected_value": {
            "type": "number",
            "title": "Expected Value"
          },
          "visit_distribution": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Visit Distribution"
          },
          "avg_reward_distribution": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Avg Reward Distribution"
          }
        },
        "type": "object",
        "required": [
          "action",
          "expected_value"
        ],
        "title": "MCTSResult"
      },
      "ScanRequest": {
        "properties": {
          "pairs": {
            "items": {
              "$ref": "#/components/schemas/AssetPair"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Pairs"
          },
          "mcts_iterations": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 100.0,
            "title": "Mcts Iterations",
            "default": 800
          },
          "source": {
            "$ref": "#/components/schemas/SignalSource",
            "default": "KALMAN_MCTS"
          }
        },
        "type": "object",
        "required": [
          "pairs"
        ],
        "title": "ScanRequest"
      },
      "ScanResponse": {
        "properties": {
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/TradeSignal"
            },
            "type": "array",
            "title": "Signals"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "scan_id",
          "signals"
        ],
        "title": "ScanResponse"
      },
      "SignalSource": {
        "type": "string",
        "enum": [
          "KALMAN_MCTS",
          "COINTEGRATION",
          "BACKTEST"
        ],
        "title": "SignalSource"
      },
      "TradeSignal": {
        "properties": {
          "pair": {
            "type": "string",
            "title": "Pair"
          },
          "action": {
            "$ref": "#/components/schemas/ActionType"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "kalman": {
            "$ref": "#/components/schemas/KalmanState"
          },
          "egarch": {
            "$ref": "#/components/schemas/EGARCHResult"
          },
          "mcts": {
            "$ref": "#/components/schemas/MCTSResult"
          },
          "source": {
            "$ref": "#/components/schemas/SignalSource"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "pair",
          "action",
          "confidence",
          "kalman",
          "egarch",
          "mcts",
          "source"
        ],
        "title": "TradeSignal"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VolatilityRegime": {
        "type": "string",
        "enum": [
          "LOW",
          "NORMAL",
          "ELEVATED",
          "CRISIS"
        ],
        "title": "VolatilityRegime"
      }
    }
  }
}